test(diagnostics): isolate visual report coverage
