+++ tests/electron/diagnostics.spec.ts
@@ -626,132 +626,6 @@ test.describe("diagnostics", () => {
     }
   });
 
-  test("refuses visual screenshot bytes without consent before export", async () => {
-    const fixture = await launchOffline("gw-visual-consent-boundary-e2e-");
-    const saveRoot = await mkdtemp(path.join(tmpdir(), "gw-visual-consent-"));
-    const target = path.join(saveRoot, "must-not-exist.zip");
-    try {
-      const outcome = await fixture.app.evaluate(
-        async ({ app: electronApp, BrowserWindow }, args) => {
-          const load = process
-            .getBuiltinModule("node:module")
-            .createRequire(args.modulePath);
-          const diagnostics = load(args.modulePath);
-          const { DEFAULT_SETTINGS } = load(args.contractsPath);
-          const { windowRegistry } = load(args.registryPath);
-          const owner = BrowserWindow.getAllWindows()[0];
-          const ownerId = owner
-            ? windowRegistry.diagnosticOwnerForWindow(owner)
-            : null;
-          if (ownerId === null) throw new Error("diagnostics owner is unavailable");
-          try {
-            await diagnostics.exportDiagnosticsZip(args.target, {
-              appVersion: electronApp.getVersion(),
-              diagnosticOwnerId: ownerId,
-              includePreviousSession: false,
-              electronVersions: { electron: process.versions.electron },
-              settings: DEFAULT_SETTINGS,
-              visualProblem: {
-                rendererOutcome: "completed",
-                gameWindowCount: 1,
-                screenshotRequested: false,
-                screenshotPng: Uint8Array.from([
-                  137, 80, 78, 71, 13, 10, 26, 10,
-                ]),
-              },
-            });
-            return "accepted";
-          } catch {
-            return "refused";
-          }
-        },
-        {
-          modulePath: path.join(root, "build/main/diagnostics.js"),
-          contractsPath: path.join(root, "build/shared/contracts.js"),
-          registryPath: path.join(root, "build/main/window-registry.js"),
-          target,
-        },
-      );
-      expect(outcome).toBe("refused");
-      expect(await stat(target).catch(() => null)).toBeNull();
-    } finally {
-      await closeOffline(fixture);
-      await rm(saveRoot, { recursive: true, force: true });
-    }
-  });
-
-  test("exports an owner-bound visual report with a consented screenshot", async () => {
-    const fixture = await launchOffline("gw-visual-problem-report-e2e-");
-    const saveRoot = await mkdtemp(path.join(tmpdir(), "gw-visual-report-"));
-    const target = path.join(saveRoot, "visual-problem.zip");
-    const extracted = path.join(saveRoot, "extracted");
-    try {
-      const { app, page } = fixture;
-      await app.evaluate(({ dialog }, filePath) => {
-        dialog.showMessageBox = async (...args: unknown[]) => {
-          const options = args.at(-1);
-          if (
-            !options
-            || typeof options !== "object"
-            || !("defaultId" in options)
-            || options.defaultId !== 1
-          ) {
-            throw new Error("visual screenshot consent must be opt-in");
-          }
-          return {
-            response: 0,
-            checkboxChecked: false,
-          };
-        };
-        dialog.showSaveDialog = async () => ({ canceled: false, filePath });
-      }, target);
-      await page.evaluate(() => {
-        window.__diagnosticExportReleasedInput = false;
-        window.addEventListener("gw:input-reset", () => {
-          window.__diagnosticExportReleasedInput = true;
-        });
-      });
-
-      await clickMenu(app, "report-visual-problem");
-      await expect.poll(async () =>
-        (await stat(target).catch(() => null))?.size ?? 0,
-      ).toBeGreaterThan(0);
-      expect(await page.evaluate(() => window.__diagnosticExportReleasedInput))
-        .toBe(true);
-
-      await execFileAsync("ditto", ["-x", "-k", target, extracted]);
-      const manifest = JSON.parse(
-        await readFile(path.join(extracted, "manifest.json"), "utf8"),
-      );
-      expect(manifest.visualProblem).toEqual({
-        rendererOutcome: "completed",
-        gameWindowCount: 1,
-        screenshotRequested: true,
-        screenshotIncluded: true,
-        screenshotPrivacy: "player-consented-unscanned",
-      });
-      expect(manifest.includedFiles).toContain("visual-problem.png");
-      const screenshot = await readFile(
-        path.join(extracted, "visual-problem.png"),
-      );
-      expect([...screenshot.subarray(0, 8)]).toEqual([
-        137, 80, 78, 71, 13, 10, 26, 10,
-      ]);
-      expect(
-        await readFile(path.join(extracted, "events.jsonl"), "utf8"),
-      ).toContain('"name":"graphics.visualProblem"');
-
-      const validated = await execFileAsync(process.execPath, [
-        path.join(root, "build/tools/diagnostics/validate.js"),
-        target,
-      ]);
-      expect(validated.stdout).toContain("valid capture");
-    } finally {
-      await closeOffline(fixture);
-      await rm(saveRoot, { recursive: true, force: true });
-    }
-  });
-
   test("a crashed client offers reporting and recovers a sandboxed renderer", async () => {
     const fixture = await launchOffline("gw-crash-panel-e2e-");
     try {
